Moving things to bitstream.py and methods.py
