+++ bitstring/__init__.py
@@ -61,7 +61,9 @@ __author__ = "Scott Griffiths"
 
 import sys
 
-from bitstring.classes import Bits, BitArray, BitStream, ConstBitStream, pack, bytealigned, lsb0, _MyModuleType
+from bitstring.classes import Bits, BitArray, bytealigned, lsb0, _MyModuleType
+from bitstring.bitstream import ConstBitStream, BitStream
+from bitstring.methods import pack
 from bitstring.bitstring_array import Array
 from bitstring.exceptions import Error, ReadError, InterpretError, ByteAlignError, CreationError
