+++ bitstring/classes.py
@@ -3314,573 +3314,6 @@ class BitArray(Bits):
     o = oct
 
 
-class ConstBitStream(Bits):
-    """A container or stream holding an immutable sequence of bits.
-
-    For a mutable container use the BitStream class instead.
-
-    Methods inherited from Bits:
-
-    all() -- Check if all specified bits are set to 1 or 0.
-    any() -- Check if any of specified bits are set to 1 or 0.
-    copy() -- Return a copy of the bitstring.
-    count() -- Count the number of bits set to 1 or 0.
-    cut() -- Create generator of constant sized chunks.
-    endswith() -- Return whether the bitstring ends with a sub-string.
-    find() -- Find a sub-bitstring in the current bitstring.
-    findall() -- Find all occurrences of a sub-bitstring in the current bitstring.
-    join() -- Join bitstrings together using current bitstring.
-    pp() -- Pretty print the bitstring.
-    rfind() -- Seek backwards to find a sub-bitstring.
-    split() -- Create generator of chunks split by a delimiter.
-    startswith() -- Return whether the bitstring starts with a sub-bitstring.
-    tobitarray() -- Return bitstring as a bitarray from the bitarray package.
-    tobytes() -- Return bitstring as bytes, padding if needed.
-    tofile() -- Write bitstring to file, padding if needed.
-    unpack() -- Interpret bits using format string.
-
-    Other methods:
-
-    bytealign() -- Align to next byte boundary.
-    peek() -- Peek at and interpret next bits as a single item.
-    peeklist() -- Peek at and interpret next bits as a list of items.
-    read() -- Read and interpret next bits as a single item.
-    readlist() -- Read and interpret next bits as a list of items.
-    readto() -- Read up to and including next occurrence of a bitstring.
-
-    Special methods:
-
-    Also available are the operators [], ==, !=, +, *, ~, <<, >>, &, |, ^.
-
-    Properties:
-
-    bin -- The bitstring as a binary string.
-    hex -- The bitstring as a hexadecimal string.
-    oct -- The bitstring as an octal string.
-    bytes -- The bitstring as a bytes object.
-    int -- Interpret as a two's complement signed integer.
-    uint -- Interpret as a two's complement unsigned integer.
-    float / floatbe -- Interpret as a big-endian floating point number.
-    bool -- For single bit bitstrings, interpret as True or False.
-    se -- Interpret as a signed exponential-Golomb code.
-    ue -- Interpret as an unsigned exponential-Golomb code.
-    sie -- Interpret as a signed interleaved exponential-Golomb code.
-    uie -- Interpret as an unsigned interleaved exponential-Golomb code.
-    floatle -- Interpret as a little-endian floating point number.
-    floatne -- Interpret as a native-endian floating point number.
-    bfloat / bfloatbe -- Interpret as a big-endian 16-bit bfloat type.
-    bfloatle -- Interpret as a little-endian 16-bit bfloat type.
-    bfloatne -- Interpret as a native-endian 16-bit bfloat type.
-    intbe -- Interpret as a big-endian signed integer.
-    intle -- Interpret as a little-endian signed integer.
-    intne -- Interpret as a native-endian signed integer.
-    uintbe -- Interpret as a big-endian unsigned integer.
-    uintle -- Interpret as a little-endian unsigned integer.
-    uintne -- Interpret as a native-endian unsigned integer.
-
-    len -- Length of the bitstring in bits.
-    pos -- The current bit position in the bitstring.
-    """
-
-    __slots__ = ()
-
-    def __init__(self, auto: Optional[BitsType] = None, length: Optional[int] = None,
-                 offset: Optional[int] = None, pos: int = 0, **kwargs) -> None:
-        """Either specify an 'auto' initialiser:
-        auto -- a string of comma separated tokens, an integer, a file object,
-                a bytearray, a boolean iterable or another bitstring.
-
-        Or initialise via **kwargs with one (and only one) of:
-        bin -- binary string representation, e.g. '0b001010'.
-        hex -- hexadecimal string representation, e.g. '0x2ef'
-        oct -- octal string representation, e.g. '0o777'.
-        bytes -- raw data as a bytes object, for example read from a binary file.
-        int -- a signed integer.
-        uint -- an unsigned integer.
-        float / floatbe -- a big-endian floating point number.
-        bool -- a boolean (True or False).
-        se -- a signed exponential-Golomb code.
-        ue -- an unsigned exponential-Golomb code.
-        sie -- a signed interleaved exponential-Golomb code.
-        uie -- an unsigned interleaved exponential-Golomb code.
-        floatle -- a little-endian floating point number.
-        floatne -- a native-endian floating point number.
-        bfloat / bfloatbe - a big-endian bfloat format 16-bit floating point number.
-        bfloatle -- a little-endian bfloat format 16-bit floating point number.
-        bfloatne -- a native-endian bfloat format 16-bit floating point number.
-        intbe -- a signed big-endian whole byte integer.
-        intle -- a signed little-endian whole byte integer.
-        intne -- a signed native-endian whole byte integer.
-        uintbe -- an unsigned big-endian whole byte integer.
-        uintle -- an unsigned little-endian whole byte integer.
-        uintne -- an unsigned native-endian whole byte integer.
-        filename -- the path of a file which will be opened in binary read-only mode.
-
-        Other keyword arguments:
-        length -- length of the bitstring in bits, if needed and appropriate.
-                  It must be supplied for all integer and float initialisers.
-        offset -- bit offset to the data. These offset bits are
-                  ignored and this is mainly intended for use when
-                  initialising using 'bytes' or 'filename'.
-        pos -- Initial bit position, defaults to 0.
-
-        """
-        if pos < 0:
-            pos += len(self._bitstore)
-        if pos < 0 or pos > len(self._bitstore):
-            raise CreationError(f"Cannot set pos to {pos} when length is {len(self._bitstore)}.")
-        self._pos = pos
-        self._bitstore.immutable = True
-
-    def _setbytepos(self, bytepos: int) -> None:
-        """Move to absolute byte-aligned position in stream."""
-        self._setbitpos(bytepos * 8)
-
-    def _getbytepos(self) -> int:
-        """Return the current position in the stream in bytes. Must be byte aligned."""
-        if self._pos % 8:
-            raise ByteAlignError("Not byte aligned when using bytepos property.")
-        return self._pos // 8
-
-    def _setbitpos(self, pos: int) -> None:
-        """Move to absolute position bit in bitstream."""
-        if pos < 0:
-            raise ValueError("Bit position cannot be negative.")
-        if pos > self.len:
-            raise ValueError("Cannot seek past the end of the data.")
-        self._pos = pos
-
-    def _getbitpos(self) -> int:
-        """Return the current position in the stream in bits."""
-        return self._pos
-
-    def _clear(self) -> None:
-        Bits._clear(self)
-        self._pos = 0
-
-    def __copy__(self) -> ConstBitStream:
-        """Return a new copy of the ConstBitStream for the copy module."""
-        # Note that if you want a new copy (different ID), use _copy instead.
-        # The copy can use the same datastore as it's immutable.
-        s = ConstBitStream()
-        s._bitstore = self._bitstore
-        # Reset the bit position, don't copy it.
-        s._pos = 0
-        return s
-
-    def __add__(self, bs: BitsType) -> Bits:
-        """Concatenate bitstrings and return new bitstring.
-
-        bs -- the bitstring to append.
-
-        """
-        s = Bits.__add__(self, bs)
-        s._pos = 0
-        return s
-
-    @overload
-    def read(self, fmt: int) -> Bits:
-        ...
-
-    @overload
-    def read(self, fmt: str) -> Any:
-        ...
-
-    def read(self, fmt: Union[int, str]) -> Union[int, float, str, Bits, bool, bytes, None]:
-        """Interpret next bits according to the format string and return result.
-
-        fmt -- Token string describing how to interpret the next bits.
-
-        Token examples: 'int:12'    : 12 bits as a signed integer
-                        'uint:8'    : 8 bits as an unsigned integer
-                        'float:64'  : 8 bytes as a big-endian float
-                        'intbe:16'  : 2 bytes as a big-endian signed integer
-                        'uintbe:16' : 2 bytes as a big-endian unsigned integer
-                        'intle:32'  : 4 bytes as a little-endian signed integer
-                        'uintle:32' : 4 bytes as a little-endian unsigned integer
-                        'floatle:64': 8 bytes as a little-endian float
-                        'intne:24'  : 3 bytes as a native-endian signed integer
-                        'uintne:24' : 3 bytes as a native-endian unsigned integer
-                        'floatne:32': 4 bytes as a native-endian float
-                        'hex:80'    : 80 bits as a hex string
-                        'oct:9'     : 9 bits as an octal string
-                        'bin:1'     : single bit binary string
-                        'ue'        : next bits as unsigned exp-Golomb code
-                        'se'        : next bits as signed exp-Golomb code
-                        'uie'       : next bits as unsigned interleaved exp-Golomb code
-                        'sie'       : next bits as signed interleaved exp-Golomb code
-                        'bits:5'    : 5 bits as a bitstring
-                        'bytes:10'  : 10 bytes as a bytes object
-                        'bool'      : 1 bit as a bool
-                        'pad:3'     : 3 bits of padding to ignore - returns None
-
-        fmt may also be an integer, which will be treated like the 'bits' token.
-
-        The position in the bitstring is advanced to after the read items.
-
-        Raises ReadError if not enough bits are available.
-        Raises ValueError if the format is not understood.
-
-        """
-        if isinstance(fmt, int):
-            if fmt < 0:
-                raise ValueError("Cannot read negative amount.")
-            if fmt > self.len - self._pos:
-                raise ReadError(f"Cannot read {fmt} bits, only {self.len - self._pos} available.")
-            bs = self._slice(self._pos, self._pos + fmt)
-            self._pos += fmt
-            return bs
-        p = self._pos
-        _, token = tokenparser(fmt)
-        if len(token) != 1:
-            self._pos = p
-            raise ValueError(f"Format string should be a single token, not {len(token)} "
-                             "tokens - use readlist() instead.")
-        name, length, _ = token[0]
-        if length is None:
-            length = self.len - self._pos
-        value, self._pos = self._readtoken(name, self._pos, length)
-        return value
-
-    def readlist(self, fmt: Union[str, List[Union[int, str]]], **kwargs) \
-            -> List[Union[float, int, str, None, Bits]]:
-        """Interpret next bits according to format string(s) and return list.
-
-        fmt -- A single string or list of strings with comma separated tokens
-               describing how to interpret the next bits in the bitstring. Items
-               can also be integers, for reading new bitstring of the given length.
-        kwargs -- A dictionary or keyword-value pairs - the keywords used in the
-                  format string will be replaced with their given value.
-
-        The position in the bitstring is advanced to after the read items.
-
-        Raises ReadError is not enough bits are available.
-        Raises ValueError if the format is not understood.
-
-        See the docstring for 'read' for token examples. 'pad' tokens are skipped
-        and not added to the returned list.
-
-        >>> h, b1, b2 = s.readlist('hex:20, bin:5, bin:3')
-        >>> i, bs1, bs2 = s.readlist(['uint:12', 10, 10])
-
-        """
-        value, self._pos = self._readlist(fmt, self._pos, **kwargs)
-        return value
-
-    def readto(self, bs: BitsType, bytealigned: Optional[bool] = None) -> Bits:
-        """Read up to and including next occurrence of bs and return result.
-
-        bs -- The bitstring to find. An integer is not permitted.
-        bytealigned -- If True the bitstring will only be
-                       found on byte boundaries.
-
-        Raises ValueError if bs is empty.
-        Raises ReadError if bs is not found.
-
-        """
-        if isinstance(bs, int):
-            raise ValueError("Integers cannot be searched for")
-        bs = Bits(bs)
-        oldpos = self._pos
-        p = self.find(bs, self._pos, bytealigned=bytealigned)
-        if not p:
-            raise ReadError("Substring not found")
-        self._pos += bs.len
-        return self._slice(oldpos, self._pos)
-
-    @overload
-    def peek(self, fmt: int) -> Bits:
-        ...
-
-    @overload
-    def peek(self, fmt: str) -> Any:
-        ...
-
-    def peek(self, fmt: Union[int, str]) -> Union[int, float, str, Bits, bool, bytes, None]:
-        """Interpret next bits according to format string and return result.
-
-        fmt -- Token string describing how to interpret the next bits.
-
-        The position in the bitstring is not changed. If not enough bits are
-        available then all bits to the end of the bitstring will be used.
-
-        Raises ReadError if not enough bits are available.
-        Raises ValueError if the format is not understood.
-
-        See the docstring for 'read' for token examples.
-
-        """
-        pos_before = self._pos
-        value = self.read(fmt)
-        self._pos = pos_before
-        return value
-
-    def peeklist(self, fmt: Union[str, List[Union[int, str]]], **kwargs) \
-            -> List[Union[int, float, str, Bits, None]]:
-        """Interpret next bits according to format string(s) and return list.
-
-        fmt -- One or more integers or strings with comma separated tokens describing
-               how to interpret the next bits in the bitstring.
-        kwargs -- A dictionary or keyword-value pairs - the keywords used in the
-                  format string will be replaced with their given value.
-
-        The position in the bitstring is not changed. If not enough bits are
-        available then all bits to the end of the bitstring will be used.
-
-        Raises ReadError if not enough bits are available.
-        Raises ValueError if the format is not understood.
-
-        See the docstring for 'read' for token examples.
-
-        """
-        pos = self._pos
-        return_values = self.readlist(fmt, **kwargs)
-        self._pos = pos
-        return return_values
-
-    def bytealign(self) -> int:
-        """Align to next byte and return number of skipped bits.
-
-        Raises ValueError if the end of the bitstring is reached before
-        aligning to the next byte.
-
-        """
-        skipped = (8 - (self._pos % 8)) % 8
-        self.pos += skipped
-        return skipped
-
-    pos = property(_getbitpos, _setbitpos,
-                   doc="""The position in the bitstring in bits. Read and write.
-                      """)
-    bitpos = property(_getbitpos, _setbitpos,
-                      doc="""The position in the bitstring in bits. Read and write.
-                      """)
-    bytepos = property(_getbytepos, _setbytepos,
-                       doc="""The position in the bitstring in bytes. Read and write.
-                      """)
-
-
-class BitStream(BitArray, ConstBitStream):
-    """A container or stream holding a mutable sequence of bits
-
-    Subclass of the ConstBitStream and BitArray classes. Inherits all of
-    their methods.
-
-    Methods:
-
-    all() -- Check if all specified bits are set to 1 or 0.
-    any() -- Check if any of specified bits are set to 1 or 0.
-    append() -- Append a bitstring.
-    bytealign() -- Align to next byte boundary.
-    byteswap() -- Change byte endianness in-place.
-    clear() -- Remove all bits from the bitstring.
-    copy() -- Return a copy of the bitstring.
-    count() -- Count the number of bits set to 1 or 0.
-    cut() -- Create generator of constant sized chunks.
-    endswith() -- Return whether the bitstring ends with a sub-string.
-    find() -- Find a sub-bitstring in the current bitstring.
-    findall() -- Find all occurrences of a sub-bitstring in the current bitstring.
-    insert() -- Insert a bitstring.
-    invert() -- Flip bit(s) between one and zero.
-    join() -- Join bitstrings together using current bitstring.
-    overwrite() -- Overwrite a section with a new bitstring.
-    peek() -- Peek at and interpret next bits as a single item.
-    peeklist() -- Peek at and interpret next bits as a list of items.
-    pp() -- Pretty print the bitstring.
-    prepend() -- Prepend a bitstring.
-    read() -- Read and interpret next bits as a single item.
-    readlist() -- Read and interpret next bits as a list of items.
-    readto() -- Read up to and including next occurrence of a bitstring.
-    replace() -- Replace occurrences of one bitstring with another.
-    reverse() -- Reverse bits in-place.
-    rfind() -- Seek backwards to find a sub-bitstring.
-    rol() -- Rotate bits to the left.
-    ror() -- Rotate bits to the right.
-    set() -- Set bit(s) to 1 or 0.
-    split() -- Create generator of chunks split by a delimiter.
-    startswith() -- Return whether the bitstring starts with a sub-bitstring.
-    tobitarray() -- Return bitstring as a bitarray from the bitarray package.
-    tobytes() -- Return bitstring as bytes, padding if needed.
-    tofile() -- Write bitstring to file, padding if needed.
-    unpack() -- Interpret bits using format string.
-
-    Special methods:
-
-    Mutating operators are available: [], <<=, >>=, +=, *=, &=, |= and ^=
-    in addition to [], ==, !=, +, *, ~, <<, >>, &, | and ^.
-
-    Properties:
-
-    bin -- The bitstring as a binary string.
-    hex -- The bitstring as a hexadecimal string.
-    oct -- The bitstring as an octal string.
-    bytes -- The bitstring as a bytes object.
-    int -- Interpret as a two's complement signed integer.
-    uint -- Interpret as a two's complement unsigned integer.
-    float / floatbe -- Interpret as a big-endian floating point number.
-    bool -- For single bit bitstrings, interpret as True or False.
-    se -- Interpret as a signed exponential-Golomb code.
-    ue -- Interpret as an unsigned exponential-Golomb code.
-    sie -- Interpret as a signed interleaved exponential-Golomb code.
-    uie -- Interpret as an unsigned interleaved exponential-Golomb code.
-    floatle -- Interpret as a little-endian floating point number.
-    floatne -- Interpret as a native-endian floating point number.
-    bfloat / bfloatbe -- Interpret as a big-endian 16-bit bfloat type.
-    bfloatle -- Interpret as a little-endian 16-bit bfloat type.
-    bfloatne -- Interpret as a native-endian 16-bit bfloat type.
-    intbe -- Interpret as a big-endian signed integer.
-    intle -- Interpret as a little-endian signed integer.
-    intne -- Interpret as a native-endian signed integer.
-    uintbe -- Interpret as a big-endian unsigned integer.
-    uintle -- Interpret as a little-endian unsigned integer.
-    uintne -- Interpret as a native-endian unsigned integer.
-
-    len -- Length of the bitstring in bits.
-    pos -- The current bit position in the bitstring.
-    """
-
-    __slots__ = ()
-
-    def __init__(self, auto: Optional[BitsType] = None, length: Optional[int] = None,
-                 offset: Optional[int] = None, pos: int = 0, **kwargs) -> None:
-        """Either specify an 'auto' initialiser:
-        auto -- a string of comma separated tokens, an integer, a file object,
-                a bytearray, a boolean iterable or another bitstring.
-
-        Or initialise via **kwargs with one (and only one) of:
-        bin -- binary string representation, e.g. '0b001010'.
-        hex -- hexadecimal string representation, e.g. '0x2ef'
-        oct -- octal string representation, e.g. '0o777'.
-        bytes -- raw data as a bytes object, for example read from a binary file.
-        int -- a signed integer.
-        uint -- an unsigned integer.
-        float / floatbe -- a big-endian floating point number.
-        bool -- a boolean (True or False).
-        se -- a signed exponential-Golomb code.
-        ue -- an unsigned exponential-Golomb code.
-        sie -- a signed interleaved exponential-Golomb code.
-        uie -- an unsigned interleaved exponential-Golomb code.
-        floatle -- a little-endian floating point number.
-        floatne -- a native-endian floating point number.
-        bfloat / bfloatbe - a big-endian bfloat format 16-bit floating point number.
-        bfloatle -- a little-endian bfloat format 16-bit floating point number.
-        bfloatne -- a native-endian bfloat format 16-bit floating point number.
-        intbe -- a signed big-endian whole byte integer.
-        intle -- a signed little-endian whole byte integer.
-        intne -- a signed native-endian whole byte integer.
-        uintbe -- an unsigned big-endian whole byte integer.
-        uintle -- an unsigned little-endian whole byte integer.
-        uintne -- an unsigned native-endian whole byte integer.
-        filename -- the path of a file which will be opened in binary read-only mode.
-
-        Other keyword arguments:
-        length -- length of the bitstring in bits, if needed and appropriate.
-                  It must be supplied for all integer and float initialisers.
-        offset -- bit offset to the data. These offset bits are
-                  ignored and this is intended for use when
-                  initialising using 'bytes' or 'filename'.
-        pos -- Initial bit position, defaults to 0.
-
-        """
-        ConstBitStream.__init__(self, auto, length, offset, pos, **kwargs)
-        if self._bitstore.immutable:
-            self._bitstore = self._bitstore.copy()
-            self._bitstore.immutable = False
-
-    def __copy__(self) -> BitStream:
-        """Return a new copy of the BitStream."""
-        s_copy = BitStream()
-        s_copy._pos = 0
-        s_copy._bitstore = self._bitstore.copy()
-        return s_copy
-
-    def prepend(self, bs: BitsType) -> None:
-        """Prepend a bitstring to the current bitstring.
-
-        bs -- The bitstring to prepend.
-
-        """
-        bs = Bits(bs)
-        super().prepend(bs)
-        self._pos += bs.len
-
-
-def pack(fmt: Union[str, List[str]], *values, **kwargs) -> BitStream:
-    """Pack the values according to the format string and return a new BitStream.
-
-    fmt -- A single string or a list of strings with comma separated tokens
-           describing how to create the BitStream.
-    values -- Zero or more values to pack according to the format.
-    kwargs -- A dictionary or keyword-value pairs - the keywords used in the
-              format string will be replaced with their given value.
-
-    Token examples: 'int:12'    : 12 bits as a signed integer
-                    'uint:8'    : 8 bits as an unsigned integer
-                    'float:64'  : 8 bytes as a big-endian float
-                    'intbe:16'  : 2 bytes as a big-endian signed integer
-                    'uintbe:16' : 2 bytes as a big-endian unsigned integer
-                    'intle:32'  : 4 bytes as a little-endian signed integer
-                    'uintle:32' : 4 bytes as a little-endian unsigned integer
-                    'floatle:64': 8 bytes as a little-endian float
-                    'intne:24'  : 3 bytes as a native-endian signed integer
-                    'uintne:24' : 3 bytes as a native-endian unsigned integer
-                    'floatne:32': 4 bytes as a native-endian float
-                    'hex:80'    : 80 bits as a hex string
-                    'oct:9'     : 9 bits as an octal string
-                    'bin:1'     : single bit binary string
-                    'ue' / 'uie': next bits as unsigned exp-Golomb code
-                    'se' / 'sie': next bits as signed exp-Golomb code
-                    'bits:5'    : 5 bits as a bitstring object
-                    'bytes:10'  : 10 bytes as a bytes object
-                    'bool'      : 1 bit as a bool
-                    'pad:3'     : 3 zero bits as padding
-
-    >>> s = pack('uint:12, bits', 100, '0xffe')
-    >>> t = pack(['bits', 'bin:3'], s, '111')
-    >>> u = pack('uint:8=a, uint:8=b, uint:55=a', a=6, b=44)
-
-    """
-    tokens = []
-    if isinstance(fmt, str):
-        fmt = [fmt]
-    try:
-        for f_item in fmt:
-            _, tkns = tokenparser(f_item, tuple(sorted(kwargs.keys())))
-            tokens.extend(tkns)
-    except ValueError as e:
-        raise CreationError(*e.args)
-    value_iter = iter(values)
-    s = BitStream()
-    try:
-        for name, length, value in tokens:
-            # If the value is in the kwd dictionary then it takes precedence.
-            if value in kwargs:
-                value = kwargs[value]
-            # If the length is in the kwd dictionary then use that too.
-            if length in kwargs:
-                length = kwargs[length]
-            # Also if we just have a dictionary name then we want to use it
-            if name in kwargs and length is None and value is None:
-                s._append(kwargs[name])
-                continue
-            if length is not None:
-                length = int(length)
-            if value is None and name != 'pad':
-                # Take the next value from the ones provided
-                value = next(value_iter)
-            s._append(BitStream._init_with_token(name, length, value))
-    except StopIteration:
-        raise CreationError(f"Not enough parameters present to pack according to the "
-                            f"format. {len(tokens)} values are needed.")
-    try:
-        next(value_iter)
-    except StopIteration:
-        # Good, we've used up all the *values.
-        return s
-    raise CreationError(
-        f"Too many parameters present to pack according to the format. Only {len(tokens)} values were expected.")
-
-
 # Whether to label the Least Significant Bit as bit 0. Default is False.
 
 def _switch_lsb0_methods(lsb0: bool) -> None:
@@ -3894,7 +3327,4 @@ def _switch_lsb0_methods(lsb0: bool) -> None:
 # Initialise the default behaviour
 _switch_lsb0_methods(False)
 
-__all__ = ['ConstBitStream', 'BitStream', 'BitArray',
-           'Bits', 'pack', 'Error', 'ReadError', 'InterpretError',
-           'ByteAlignError', 'CreationError', 'bytealigned', 'lsb0']
